requirements.txt
streamlit
openai
langchain==0.0.342
streamlit-webrtc
google-generativeai==0.3.2
pydub
pillow
python-dotenv
opencv-python-headless==4.8.1.78
